bugfix-library-page
Делает кнопку Watched активной
